A03_DVC_Pipeline/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asyncssh==2.14.2
atpublic==4.0
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.8
cryptography==42.0.5
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.1.6
dulwich==0.21.7
dvc==3.48.4
dvc-data==3.14.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.1
dvc-studio-client==0.20.0
dvc-task==0.4.0
entrypoints==0.4
filelock==3.13.3
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.4.1
fsspec==2024.3.1
funcy==2.0
gitdb==4.0.11
GitPython==3.1.42
grandalf==0.8
gto==1.7.1
hydra-core==1.3.2
idna==3.6
iterative-telemetry==0.0.8
joblib==1.3.2
kombu==5.3.5
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
orjson==3.9.15
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
platformdirs==3.11.0
prompt-toolkit==3.0.43
psutil==5.9.8
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydot==2.0.0
pygit2==1.14.1
Pygments==2.17.2
pygtrie==2.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.4.1.post1
scipy==1.12.0
scmrepo==3.3.1
semver==3.0.2
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sqltrie==0.11.0
tabulate==0.9.0
threadpoolctl==3.4.0
tomlkit==0.12.4
tqdm==4.66.2
typer==0.10.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
voluptuous==0.14.2
wcwidth==0.2.13
wget==3.2
yarl==1.9.4
zc.lockfile==3.0.post1
